feat: Nix development shell
We add a simple nix development shell, based on pyproject.nix, as well
as some notes in the readme about it.

There are also some minor fixes and changes; the most notable being set
set the `__version__` field optionally, if the package is actually
installed, which if it is done using pip it well be, but via
pyproject.nix, it will not be.